ZMQ-83: "Resource temporarily unavailable" when running on 266MHz box
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@527 7451a71c-b6c8-49cc-8254-141fc5361860

+++ zmq/tcp_socket.hpp
@@ -51,13 +51,16 @@ namespace zmq
             return s;
         }
 
-        //  Writes data to the socket. Returns the number
-        //  of bytes actually written.
-        size_t write (const void *data, size_t size);
+        //  Writes data to the socket. Returns the number of bytes actually
+        //  written (even zero is to be considered to be a success). In case
+        //  of orderly shutdown by the other peer -1 is returned.
+        int write (const void *data, int size);
 
         //  Reads data from the socket (up to 'size' bytes). Returns the number
-        //  of bytes actually read.
-        size_t read (void *data, size_t size);
+        //  of bytes actually read (even zero is to be considered to be
+        //  a success). In case of orderly shutdown by the other peer -1 is
+        //  returned.
+        int read (void *data, int size);
 
     private:
